Simplify data model usage by keeping it in Kotlin files

+++ app/src/main/java/io/aether/android/screens/device/settings/FabricsViewModel.kt
@@ -14,7 +14,6 @@ import io.aether.android.matter.FabricId
 import io.aether.android.matter.NodeId
 import io.aether.android.matter.VendorId
 import io.aether.android.matter.toFabricId
-import io.aether.android.matter.toLong
 import io.aether.android.matter.toNodeId
 import io.aether.android.matter.toVendorId
 import javax.inject.Inject
@@ -72,8 +71,7 @@ constructor(
       val currentState = _uiState.value
       _uiState.value = UiState.Loading
       try {
-        val deviceCurrentFabricIndex =
-            clustersHelper.readCurrentFabricIndexAttribute(nodeId.toLong())
+        val deviceCurrentFabricIndex = clustersHelper.readCurrentFabricIndexAttribute(nodeId)
         val controllerFabricIndex = chipClient.chipDeviceController.getFabricIndex()
         val currentFabricIndex = deviceCurrentFabricIndex ?: controllerFabricIndex
         if (fabricIndex == currentFabricIndex) {
@@ -81,7 +79,7 @@ constructor(
           _uiState.value = currentState
           return@launch
         }
-        clustersHelper.removeFabric(nodeId.toLong(), fabricIndex)
+        clustersHelper.removeFabric(nodeId, fabricIndex)
         _uiState.value = refreshFabrics(nodeId, currentState)
       } catch (e: Exception) {
         Timber.e(e, "removeFabric failed")
@@ -94,15 +92,14 @@ constructor(
 
   private suspend fun refreshFabrics(nodeId: NodeId, fallbackState: UiState): UiState {
     return try {
-      devicesRepository.getDeviceByNodeId(nodeId.toLong())
-      val fabrics = clustersHelper.readFabricsAttribute(nodeId.toLong())
-      val nocs = clustersHelper.readNOCsAttribute(nodeId.toLong())
+      devicesRepository.getDeviceByNodeId(nodeId)
+      val fabrics = clustersHelper.readFabricsAttribute(nodeId)
+      val nocs = clustersHelper.readNOCsAttribute(nodeId)
       if (fabrics == null || nocs == null) {
         if (fallbackState is UiState.Loaded) fallbackState
         else UiState.Error(R.string.controllers_offline)
       } else {
-        val deviceCurrentFabricIndex =
-            clustersHelper.readCurrentFabricIndexAttribute(nodeId.toLong())
+        val deviceCurrentFabricIndex = clustersHelper.readCurrentFabricIndexAttribute(nodeId)
         val controllerFabricIndex = chipClient.chipDeviceController.getFabricIndex()
         val currentFabricIndex = deviceCurrentFabricIndex ?: controllerFabricIndex
         val fabricsByIndex = fabrics.associateBy { it.fabricIndex }
